requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
networkx==3.4.2
numpy==2.1.3
packaging==24.2
scikit-fuzzy==0.5.0
scipy==1.14.1
sqlparse==0.5.1
tzdata==2024.2
